Add onboarding feature with view model, view, and widgets
This commit adds the onboarding feature including the view model, view, and
various widgets for displaying onboarding screens. It also updates imports and
removes the old onboarding screen.
