+++ lib/Features/onBoarding/Presentation/View_model/boarding_model.dart
@@ -1,0 +1,10 @@
+class BoardingModel {
+  String title;
+  String image;
+  String body;
+  BoardingModel({
+    required this.image,
+    required this.title,
+    required this.body,
+  });
+}
