+++ pubspec.yaml
@@ -33,6 +33,7 @@ dependencies:
     sdk: flutter
   flutter_svg: ^2.0.10+1
   google_fonts: ^6.2.1
+  smooth_page_indicator: ^1.1.0
 
 dev_dependencies:
   appbar_animated: ^0.0.3
